# Codigo/backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "dtiflix-backend",
    "description": "Full stack application built with Next.js + TypeScript and Sequelize, integrating the OMDb API to search movies, manage favorites, write reviews, and display personalized statistics in a dashboard. Includes search history, average ratings, favorite genres/decades, and a responsive UI.",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "termsOfService": "http://localhost:5000/api/",
    "contact": {
      "name": "Davi Cândido",
      "email": "davicandidopucminas@gmail.com",
      "phone": "+55 (31) 97306-7259"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/",
      "description": "Ambiente de desenvolvimento"
    },
    {
      "url": "http://localhost:5000/api/",
      "description": "Ambiente de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "imdbID": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "decade": {
            "type": "integer"
          },
          "plot": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "rated": {
            "type": "string"
          },
          "runtime": {
            "type": "string"
          },
          "director": {
            "type": "string"
          },
          "actors": {
            "type": "string"
          },
          "poster": {
            "type": "string"
          },
          "imdbRating": {
            "type": "number"
          },
          "ratings": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "type": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/movies/search/{imdbID}/{queryTerm}": {
      "get": {
        "summary": "Busca filme por IMDb ID com registro de histórico de pesquisa",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "imdbID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "IMDb ID do filme"
          },
          {
            "in": "path",
            "name": "queryTerm",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Termo de pesquisa usado pelo usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Filme encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      }
    },
    "/movies/title/{title}": {
      "get": {
        "summary": "Busca filmes pelo título",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Título do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/genre/{genre}": {
      "get": {
        "summary": "Busca filmes por gênero",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "genre",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Gênero do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/year/{year}": {
      "get": {
        "summary": "Busca filmes por ano",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Ano do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/decade/{decade}": {
      "get": {
        "summary": "Busca filmes por década",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "decade",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "Década do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies": {
      "get": {
        "summary": "Lista todos os filmes",
        "tags": ["Movies"],
        "responses": {
          "200": {
            "description": "Lista de filmes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo filme",
        "tags": ["Movies"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Filme criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "summary": "Busca filme pelo ID",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Filme encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza filme",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Deleta filme",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Filme deletado"
          }
        }
      }
    },
    "/movies/{movieId}/reviews": {
      "get": {
        "summary": "Lista avaliações (reviews) de um filme",
        "tags": ["Movies"],
        "parameters": [
          {
            "in": "path",
            "name": "movieId",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID do filme"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de avaliações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}